+++ apps/api/src/app/app.module.ts
@@ -7,10 +7,7 @@ import { MongooseModule } from '@nestjs/mongoose'
 import { GraphQLError } from 'graphql'
 import { join } from 'path'
 
-import { SearchModule } from '@api/search/search.module'
-
 import { AuthModule } from '../auth/auth.module'
-import { ClientLoggingModule } from '../clientlogging/clientlogging.module'
 import { CommonModule } from '../common/common.module'
 import { GraphQLExpressContext } from '../common/types/context.type'
 import { ComputationModule } from '../computation/computation.module'
@@ -80,10 +77,8 @@ import { AppService } from './app.service'
     UserModule,
     AuthModule,
     ReviewModule,
-    ClientLoggingModule,
     OverrideModule,
     ComputationModule,
-    SearchModule,
   ],
   controllers: [AppController],
   providers: [AppService],