+++ apps/api/CHANGELOG.md
@@ -1,5 +1,27 @@
 # api
 
+## 1.13.0-beta.1
+
+### Minor Changes
+
+- 5eb12f2: refactor: remove client logging
+
+### Patch Changes
+
+- Updated dependencies [5eb12f2]
+  - @cgr/schema@1.7.0-beta.1
+
+## 1.13.0-beta.0
+
+### Minor Changes
+
+- 18f4e95: refactor: remove opensearch
+
+### Patch Changes
+
+- Updated dependencies [18f4e95]
+  - @cgr/schema@1.7.0-beta.0
+
 ## 1.12.0
 
 ### Minor Changes