+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cgr/schema
 
+## 1.7.0-beta.1
+
+### Minor Changes
+
+- 5eb12f2: refactor: remove client logging
+
+## 1.7.0-beta.0
+
+### Minor Changes
+
+- 18f4e95: refactor: remove opensearch
+
 ## 1.6.0
 
 ### Minor Changes